Restructuring
CSS was restructured with hardware to make the "contactor" something replicable for future projects.

In the counter, we've included accordion items for the specific region, this means that the user now has the freedom to hide exclusions that he eventually doesn't want to see.

An entire typography was styled for a new layout concept.

A new counter has been added but is still under development.